chore(tests): lots of tests & refactor
Gist -- this commit gets everything baselined against the domain-specific repositories (not yet organized) with *some* test coverage. The test coverage highlights a failing test -- specifically, there's an issue projecting w/ nested automappings. It's unclear if this issue is because we're trying to project to record types or whether it's simply a result of the nested covnersion going beyond the limitations of linq->sql projection